neo: extract reusable daemon service system into neo.
This is a fairly large refactoring that extracts much of the service
model into neo so it can be used elsewhere. This makes services to be
owned by the application, which is also a service. You can locate peer
services to communicate with each other.

Basic logging is simply and can be done by setting :loggind-enabled
property of the application.